@@ -63,7 +63,7 @@ dev-dependencies = [
 [tool.ruff]
 src = ["src"]
 exclude = []
-target-version = "py39"
+target-version = "py310"
 line-length = 120
 lint.select = [
   "A",
@@ -144,7 +144,7 @@ skip_empty = true
 sort = "cover"
 
 [tool.mypy]
-python_version = "3.9"
+python_version = "3.10"
 warn_return_any = true
 warn_unused_configs = true
 ignore_missing_imports = true
